+++ app/View/Livewire/Fragments.php
@@ -16,6 +16,7 @@ class Fragments extends Component
     public FragmentsOwner $owner;
     private Collection $fragments;
     private array $allowedFragments;
+    private array $sharedFragments;
 
     public function mount(FragmentsOwner $owner)
     {
@@ -27,32 +28,48 @@ class Fragments extends Component
     public function render()
     {
         return view('chief::components.fragments', [
-            'fragments' => $this->fragments,
+            'fragments'        => $this->fragments,
             'allowedFragments' => $this->allowedFragments,
-            'manager' => app(Registry::class)->manager($this->owner::managedModelKey()),
+            'sharedFragments'  => $this->sharedFragments,
+            'manager'          => app(Registry::class)->manager($this->owner::managedModelKey()),
         ]);
     }
 
     public function reload()
     {
-        // Current fragments
         $this->fragments = app(FragmentRepository::class)->getByOwner($this->owner->ownerModel())->map(function (Fragmentable $model) {
             return [
-                'model' => $model,
+                'model'   => $model,
                 'manager' => app(Registry::class)->manager($model::managedModelKey()),
             ];
         });
 
+        $this->setAllowedFragments();
+
+        $this->emit('fragmentsReloaded');
+    }
+
+    private function setAllowedFragments()
+    {
         // Available fragments
         $this->allowedFragments = array_map(function ($fragmentableClass) {
             $modelClass = app(Registry::class)->modelClass($fragmentableClass::managedModelKey());
 
             return [
                 'manager' => app(Registry::class)->manager($fragmentableClass::managedModelKey()),
-                'model' => new $modelClass(),
+                'model'   => new $modelClass(),
             ];
         }, $this->owner->allowedFragments());
 
-        $this->emit('fragmentsReloaded');
+        $fragmentModelIds = $this->fragments->map(fn($fragment) => $fragment['model']->fragmentModel())->pluck('id')->toArray();
+
+        $this->sharedFragments = app(FragmentRepository::class)->shared()->reject(function($fragmentable) use($fragmentModelIds){
+            return in_array($fragmentable->fragmentModel()->id, $fragmentModelIds);
+        })->map(function($fragmentable){
+            return [
+                'manager' => app(Registry::class)->manager($fragmentable::managedModelKey()),
+                'model'   => $fragmentable,
+            ];
+        })->all();
     }
 }