+++ src/Fragments/Database/FragmentRepository.php
@@ -39,6 +39,13 @@ final class FragmentRepository
         );
     }
 
+    public function shared(): Collection
+    {
+        return FragmentModel::where('shared',1)->get()->map(
+            fn (FragmentModel $fragmentModel) => $this->fragmentFactory($fragmentModel)
+        );
+    }
+
     public function find(string $id): Fragmentable
     {
         return $this->fragmentFactory(FragmentModel::findOrFail($id));