+++ src/Fragments/FragmentAlreadyAdded.php
@@ -1,0 +1,9 @@
+<?php
+declare(strict_types=1);
+
+namespace Thinktomorrow\Chief\Fragments;
+
+class FragmentAlreadyAdded extends \Exception
+{
+
+}